+++ src/api_calls/version_calls.rs
@@ -1,13 +1,13 @@
 use super::super::structures::version_structs::*;
 use crate::{
     request::{request, request_rel},
-    ModrinthAPI, Result,
+    Ferinth, Result,
 };
 use bytes::Bytes;
 
-impl ModrinthAPI {
+impl Ferinth {
     /// Get the versions of mod with `mod_id`
-    pub async fn list_versions(&self, mod_id: String) -> Result<Vec<Version>> {
+    pub async fn list_versions(&self, mod_id: &str) -> Result<Vec<Version>> {
         Ok(request_rel(self, format!("/mod/{}/version", mod_id))
             .await?
             .json()
@@ -15,7 +15,7 @@ impl ModrinthAPI {
     }
 
     /// Get version with ID `version_id`
-    pub async fn get_version(&self, version_id: String) -> Result<Version> {
+    pub async fn get_version(&self, version_id: &str) -> Result<Version> {
         Ok(request_rel(self, format!("/version/{}", version_id))
             .await?
             .json()
@@ -23,7 +23,7 @@ impl ModrinthAPI {
     }
 
     /// Download `version_file`'s contents
-    pub async fn download_version_file(&self, version_file: VersionFile) -> Result<Bytes> {
-        Ok(request(self, version_file.url).await?.bytes().await?)
+    pub async fn download_version_file(&self, version_file: &VersionFile) -> Result<Bytes> {
+        Ok(request(self, &version_file.url).await?.bytes().await?)
     }
 }