+++ src/error.rs
@@ -1,9 +1,0 @@
-use thiserror::Error;
-
-pub type Result<T> = std::result::Result<T, Error>;
-
-#[derive(Error, Debug)]
-pub enum Error {
-    #[error("A HTTP(S) error occured {0}")]
-    HTTPError(#[from] reqwest::Error),
-}